+++ changelog.md
@@ -5,3 +5,7 @@
 ### 2021-06-20
 - fixed: issues with the Xception model and removed base model arch
 - added: trainer and model manager for managing the models file
+
+### 2021-06-22
+- added: data loader to load images from the file directories in a form of image tensor
+- completed data loading pipeline